requirements.txt
keras
matplotlib
numpy
Pillow
